Silence size conversion warnings under Win64 where it is safe to to so because we know the actual value ranges.
* subversion/libsvn_repos/load.c
  (parse_text_block,
   svn_repos_parse_dumpstream3): bufflen is limited by stream chunk size
* subversion/svn/mergeinfo-cmd.c
  (mergeinfo_diagram): those text snippets are short
* subversion/tests/libsvn_subr/io-test.c
  (create_test_file): size is <= 5

git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1405530 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svn/mergeinfo-cmd.c
@@ -153,7 +153,7 @@ mergeinfo_diagram(svn_client__pathrev_t *yca,
       for (row = 0; row < ROWS; row++)
         {
           if (g[row][col] && (strlen(g[row][col]) > col_width[col]))
-            col_width[col] = strlen(g[row][col]);
+            col_width[col] = (int)strlen(g[row][col]);
         }
     }
